feat: move location utils to the folder
