+++ src/extensions/hooks/adopted-probe/tsconfig.json
@@ -30,7 +30,7 @@
 	},
 	"include": [
 		"./src/**/*.ts",
-		"../../lib/*.ts",
+		"../../lib/**/*.ts",
 		"./test/**/*.ts"
 	]
 }
