+++ src/extensions/hooks/notifications-format/tsconfig.json
@@ -29,7 +29,7 @@
 	},
 	"include": [
 		"./src/**/*.ts",
-		"../../lib/*.ts",
+		"../../lib/**/*.ts",
 		"./test/**/*.ts"
 	]
 }
