+++ src/extensions/interfaces/visible-token/tsconfig.json
@@ -29,7 +29,7 @@
 	},
 	"include": [
 		"./src/**/*.ts",
-		"../../lib/*.ts",
+		"../../lib/**/*.ts",
 		"./test/**/*.ts"
 	]
 }
