+++ src/extensions/operations/check-outdated-firmware-cron-handler/tsconfig.json
@@ -29,7 +29,7 @@
 	},
 	"include": [
 		"./src/**/*.ts",
-		"../../lib/*.ts",
+		"../../lib/**/*.ts",
 		"./test/**/*.ts"
 	]
 }
